split spotlighSubmissionService with spotlighBatchService
